+++ rest/src/main/java/fi/otavanopisto/pyramus/rest/controller/CourseController.java
@@ -1,5 +1,6 @@
 package fi.otavanopisto.pyramus.rest.controller;
 
+import java.math.BigDecimal;
 import java.util.Currency;
 import java.util.Date;
 import java.util.HashSet;
@@ -458,7 +459,7 @@ public class CourseController {
   /* CourseStudent */
   
   public CourseStudent createCourseStudent(Course course, Student student, CourseEnrolmentType enrolmentType, CourseParticipationType participationType, 
-      Date enrolmentDate, Boolean lodging, CourseOptionality optionality, BillingDetails billingDetails, Double lodgingFee, Currency lodgingFeeCurrency, String organization, String additionalInfo, Room room) {
+      Date enrolmentDate, Boolean lodging, CourseOptionality optionality, BillingDetails billingDetails, BigDecimal lodgingFee, Currency lodgingFeeCurrency, String organization, String additionalInfo, Room room) {
     return courseStudentDAO.create(course, student, enrolmentType, participationType, enrolmentDate, lodging, optionality, billingDetails, organization, additionalInfo, room, lodgingFee, lodgingFeeCurrency, Boolean.FALSE);
   }
